porn91/requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiohttp-jinja2==1.2.0
aiohttp-socks==0.7.1
aiosignal==1.2.0
aioweb @ git+https://gitee.com/shadaileng/packaging.git@0cc8cd917e4cb5e51b3ed53c67ae065de585fb27
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bencoder.pyx==3.0.0
Brotli==1.0.9
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
# frozenlist==1.3.3
future==0.18.2
idna==2.10
Jinja2==3.0.2
Js2Py==0.71
lmdb==1.4.1
MarkupSafe==2.0.1
multidict==4.7.6
ping3==4.0.0
psutil==5.9.0
pycryptodome==3.14.1
pyjsparser==2.7.1
python-socks==2.0.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.24.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.18
tzdata==2022.1
tzlocal==4.2
urllib3==1.25.11
yarl==1.7.2
